Crash bug fixed
Fixed a bug in which if the player car crashed nose into another's tail, the force and torque added would not be enough to lift it off the ground. Instead, an explosion force was added, with an origin underneath the player and a radius and upwards force enough to achieve the desired effect. Frontal crashes flip the car downwards while side crashes do a lot more random stuff.